requirements.txt
Flask
pyapns
BeautifulSoup4
python-gcm
service_identity
setproctitle
